First commit, taken just at the end of Part 2 / start of Part 3.
This commit was taken just after Austin saves the scene as Level 1 but
before he shows us how to create Level 2, roughly 1 minute in to Part 3.

It contains everything from the start of the tutorial, through the
lessons on adding a player, making him move, adding an enemy and making
him move and finally adding a goal that doesn't do anything yet.