Committed at the end of Part 4.
Committed at the end of Part 4 after Austin has shown us how to create a
Die method, Die when we fall off the level, create a GameManager class,
add level changing when touching the goal, add 2 new levels, rotate the
camera and player to give a better view and keep controls working,
change the method to add force to the player, rotate the death particles
to make the blast go upwards. There is also some colour changing which
is slightly different in his version.